Flipper File Format: enhanced version
